[Android] onFirstRunCheckDone was removed
Chromium change:
https://github.com/chromium/chromium/commit/2e294c447ae4a0c153226a2372118f937b978b93

[android] Stop blocking sign-in if FRE isn't done
This requirement was introduced in crbug.com/303953 as a way to
prevent a background sign-in in a use case for EDU accounts (the
background sign-in would break some assumptions and cause a crash).
This use case should not be possible anymore, so the CL just deletes
the code.

If the issue somehow resurfaces, we can further patch
isSigninAllowed() and isSyncOptInAllowed() to check that the FRE is
complete, instead of relying on callers signaling onFirstRunCheckDone()
themselves.

Fixed: 1316369